Fix underflow checks in float8in().
We use "long double" in order to perform the underflow check in GPDB.
Lost this little change in the merge.

I don't know why we do that in GPDB, it seems like a pretty random little
behavioral difference from upstream. But let's not second guess that as
part of the merge.

+++ src/backend/utils/adt/float.c
@@ -445,7 +445,7 @@ double
 float8in_internal(char *num, char **endptr_p,
 				  const char *type_name, const char *orig_string)
 {
-	double		val;
+	long double		val;
 	char	   *endptr;
 	bool 		literal_inf = true;
